+++ sdk/core/core-http/src/util/throttlingRetryStrategy.ts
@@ -1,0 +1,4 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+export const DEFAULT_CLIENT_MAX_RETRY_COUNT = 3;