+++ sdk/core/core-http/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features Added
 
 - Added support for the `Retry-After` header on responses with status code 503, Service Unavailable.
+- Added support for multiple retries on the `ThrottlingRetryPolicy` (up to 3 by default).
 
 ### Breaking Changes