+++ sdk/core/core-rest-pipeline/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Features Added
 
 - Added support for the `Retry-After` header on responses with status code 503, Service Unavailable.
+- The `ExponentialRetryPolicy` will now ignore `503` responses if they have the `Retry-After` header.
 
 ## 1.1.0-beta.3 (2021-06-03)